Channel invitation in a separate component
